Chicken Coop Door
This code powers an Arduino UNO microcontroller to open and close a
chicken coop door based on change in light. This is perfect for opening
and closing of the door in the morning and at night. There is motor
protection included with ramping up / down the motor with sinusoidal
motion with each powering of the motor. Lastly, there is programming for
OFF / AUTO / FORWARD / REVERSE settings on a step-resistor. AUTO is used
to implement the light detection.